@@ -3,12 +3,10 @@ package com.beijing.zzu.plugindemo;
 import android.Manifest;
 import android.content.Intent;
 import android.content.pm.PackageManager;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.app.ActivityCompat;
 import android.os.Environment;
-import android.support.annotation.NonNull;
-import android.support.v4.app.ActivityCompat;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.View;
 
 public class MainActivity extends AppCompatActivity {
@@ -35,7 +33,7 @@ public class MainActivity extends AppCompatActivity {
     // 编译pluginapk工程生成pluginapk-debug.apk之后，执行adb push命令将.apk推到手机上
     //  adb push D:\Dev\GitHub\AndroidPluginDemo\pluginapk\build\outputs\apk\debug\pluginapk-debug.apk /sdcard
     @Override
-    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
+    public void onRequestPermissionsResult(int requestCode, String[] permissions, int[] grantResults) {
         super.onRequestPermissionsResult(requestCode, permissions, grantResults);
         if (requestCode == 100) {
             if (grantResults.length > 0 &&
